+++ CHANGELOG.md
@@ -3,9 +3,17 @@ odbc CHANGELOG
 
 This file is used to list changes made in each version of the odbc cookbook.
 
+0.1.2
+-----
+- [Justin Schneck] - Added freetds driver
+
+0.1.1
+-----
+- [Justin Schneck] - Changed default odbc version to 2.3.2
+
 0.1.0
 -----
-- [your_name] - Initial release of odbc
+- [Justin Schneck] - Initial release of odbc
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.